+++ packages/dataviews/src/components/dataviews-layouts/picker-grid/index.tsx
@@ -332,7 +332,7 @@ function ViewPickerGrid< Item >( {
 		isItemSelectable: () => true,
 		selection,
 		onChangeSelection,
-		isMultiselect,
+		selectionMode: isMultiselect ? 'multi' : 'single-clearable',
 		shouldSelectOnClick: true,
 	} );