+++ packages/dataviews/src/components/dataviews-layouts/picker-table/index.tsx
@@ -271,7 +271,7 @@ function ViewPickerTable< Item >( {
 		isItemSelectable: () => true,
 		selection,
 		onChangeSelection,
-		isMultiselect,
+		selectionMode: isMultiselect ? 'multi' : 'single-clearable',
 		shouldSelectOnClick: true,
 	} );