+++ packages/dataviews/src/components/dataviews-layouts/table/index.tsx
@@ -295,7 +295,7 @@ function ViewTable< Item >( {
 		isItemSelectable: ( item ) => hasAPossibleBulkAction( actions, item ),
 		selection,
 		onChangeSelection,
-		isMultiselect: true,
+		selectionMode: 'multi',
 		shouldSelectOnClick: false,
 	} );
 	const headerMenuRefs = useRef<